requirements.txt
streamlit==1.45.0
requests==2.32.3
grpcio==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.4 
python-dotenv